finish api

+++ app/Http/Controllers/PostController.php
@@ -4,9 +4,20 @@ namespace App\Http\Controllers;
 
 use App\Models\Post;
 use Illuminate\Http\Request;
+use Illuminate\Routing\Controllers\HasMiddleware;
+use Illuminate\Routing\Controllers\Middleware;
+use Illuminate\Support\Facades\Gate;
 
-class PostController extends Controller
+class PostController extends Controller implements HasMiddleware
 {
+
+    public static function middleware()
+    {
+        return [
+            new Middleware('auth:sanctum', except: ['index, show'])
+        ];
+    }
+
     /**
      * Display a listing of the resource.
      */
@@ -24,7 +35,7 @@ class PostController extends Controller
             'title' => 'required|max:255',
             'body' => 'required',
         ]);
-        $post = Post::create($fields);
+        $post = $request->user()->posts()->create($fields);
         return ['post' => $post];
     }
 
@@ -41,6 +52,8 @@ class PostController extends Controller
      */
     public function update(Request $request, Post $post)
     {
+        Gate::authorize('modufy', $post);
+
         $fields = $request->validate([
             'title' => 'required|max:255',
             'body' => 'required',
@@ -54,6 +67,8 @@ class PostController extends Controller
      */
     public function destroy(Post $post)
     {
+        Gate::authorize('modufy', $post);
+
         $post->delete();
         return ['message' => 'Post was deleted'];
     }
